Responses/comments in remark.py
Still some work to do.

I'd say there are 4 scenarios in which a user might be running the notebook:

1. Jupyter notebook
1. Spyder or other IDE
1. Command line on machine with graphics
1. Command line on machine without graphics

We need to set things up so that our code does not blow up in any of these scenarios.  (That does not necessarily mean that it has to work; just that if it does not work, the user should get clear explanations for why).

@mriduls will need to weigh in on one question, which is what will work on a machine without a GUI.  Will such a machine have iPython installed on it, and will iPython commands work?  If not, my understanding that %matplotlab inline will not work.